+++ examples/search/search.py
@@ -1,11 +1,9 @@
 from aim.sdk import select
-from aim.engine.repo import AimRepo
-from aim.ql.grammar.statement import Statement
 
 
 if __name__ == '__main__':
     res = select('loss if '
-                 'context.subset == train '
+                 'experiment == test_indicator and run.archived is True'
                  )
 
     print(res)